+++ CHANGELOG.md
@@ -17,11 +17,11 @@ reset-password`
 hash, in place of the cleartext password
 - Added the beginnings of the core/v3 API.
 - Added automatically generated tests for the core/v2 API.
-- Added suport for the `aix/ppc64` platform.
 
 ### Changed
 - `sensu-agent`'s default log level is now `info` instead of `warn`.
-- Upgrade gostatsd to 20.3.2.
+- `sensuctl command exec` now spawns the plugin directly instead of going
+  through a sub-shell.
 
 ### Fixed
 - The password verification logic when running `sensuctl user change-password`